requirements.txt
torch
torchvision
pyyaml
matplotlib